Unify BPF pointers generation
