+++ scapy/arch/common.py
@@ -10,6 +10,8 @@ Functions common to different architectures
 import socket
 from fcntl import ioctl
 import struct
+from ctypes import POINTER, Structure
+from ctypes import c_uint, c_uint32, c_ushort, c_ubyte
 
 def get_if(iff, cmd):
     """Ease SIOCGIF* ioctl calls"""
@@ -18,3 +20,37 @@ def get_if(iff, cmd):
     ifreq = ioctl(sck, cmd, struct.pack("16s16x", iff.encode("utf8")))
     sck.close()
     return ifreq
+
+class bpf_insn(Structure):
+    """"The BPF instruction data structure"""
+    _fields_ = [("code", c_ushort),
+                ("jt", c_ubyte),
+                ("jf", c_ubyte),
+                ("k", c_uint32)]
+
+
+class bpf_program(Structure):
+    """"Structure for BIOCSETF"""
+    _fields_ = [("bf_len", c_uint),
+                ("bf_insns", POINTER(bpf_insn))]
+
+def get_bpf_pointer(tcpdump_lines):
+    """Create a BPF Pointer for TCPDump filter"""
+    # Allocate BPF instructions
+    size = int(tcpdump_lines[0])
+    bpf_insn_a = bpf_insn * size
+    bip = bpf_insn_a()
+
+    # Fill the BPF instruction structures with the byte code
+    tcpdump_lines = tcpdump_lines[1:]
+    i = 0
+    for line in tcpdump_lines:
+        values = [int(v) for v in line.split()]
+        bip[i].code = c_ushort(values[0])
+        bip[i].jt = c_ubyte(values[1])
+        bip[i].jf = c_ubyte(values[2])
+        bip[i].k = c_uint(values[3])
+        i += 1
+
+    # Create the BPF program
+    return bpf_program(size, bip)
